file uploadati

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -7,11 +7,13 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Admin\Project;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Models\Admin\Technology;
 use App\Models\Admin\Type;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use PhpParser\Node\Stmt\If_;
 
+use function PHPUnit\Framework\isNull;
 
 class ProjectController extends Controller
 {
@@ -22,14 +24,16 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        $projects= Project::all();
+        $projects = Project::all();
 
         $types = Type::all();
+        $technologies = Technology::all();
 
 
         return view("admin.projects.index", [
-            "projects" => $projects,    
-            "types" => $types    
+            "projects" => $projects,
+            "types" => $types,
+            "technologies" => $technologies
         ]);
     }
 
@@ -41,8 +45,9 @@ class ProjectController extends Controller
     public function create()
     {
         $types = Type::all();
- 
-        return view("admin.projects.create", compact("types"));
+        $technologies = Technology::all();
+
+        return view("admin.projects.create", compact("types", "technologies"));
     }
 
     /**
@@ -53,34 +58,41 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
-       $data = $request->validate([
-        "name" => "required|string|max:20",
-        "description" => "required|string",
-        "cover_img" => "file",
-        "github_link" => "string",
-        "type_id" => "nullable|exists:types,id"
+        $data = $request->validate([
+            "name" => "required|string|max:20",
+            "description" => "required|string",
+            "cover_img" => "file",
+            "github_link" => "string",
+            "type_id" => "nullable|exists:types,id",
+            "technology_id" => "array|nullable|exists:technologies,id"
         ]);
 
 
-        if (key_exists("cover_img", $data)){
+        if (key_exists("cover_img", $data)) {
 
             $path = Storage::put("projects", $data["cover_img"]);
         }
- 
-       $project = Project::create([
-        ...$data,
-        //a bd vado a salvare solamente il percorso 
-        "cover_img" => $path ?? '',
-        // recuperiamo l'id dagli user cioé user_id é uguale all'utente loggato
-        "user_id" => Auth::id() 
+
+        $project = Project::create([
+            ...$data,
+            //a bd vado a salvare solamente il percorso 
+            "cover_img" => $path ?? '',
+            // recuperiamo l'id dagli user cioé user_id é uguale all'utente loggato
+            "user_id" => Auth::id()
         ]);
-
+        /*  dd($data); */
+        /* prende i dati e vede se esistono quei valori nella tabella technlologies  */
+        /* la funzione technologie è quella del model technology */
+        if ($request->has("technology_id")) {
+            /* oppure posso usare */
+            // if (key_exists("tags", $data)) {
+            $project->technologies()->attach($data["technology_id"]);
+        }
+        /* dd($data); */
         return redirect()->route("admin.projects.show", compact("project"));
-
-       
     }
-       
-   
+
+
 
     /**
      * Display the specified resource.
@@ -91,9 +103,9 @@ class ProjectController extends Controller
     public function show($id)
     {
         $project = Project::findOrFail($id);
-        $types = Type::all();
-  
-        return view("admin.projects.show", compact("project","types"));
+
+       /*  dd($project); */
+        return view("admin.projects.show", ["project" => $project]);
     }
 
     /**
@@ -104,10 +116,11 @@ class ProjectController extends Controller
      */
     public function edit($id)
     {
-        $types = Type::all();
+        $type = Type::all();
+        $technologies = Technology::all();
 
         $project = Project::findOrFail($id);
-        return view("admin.projects.edit",compact("project","types"));
+        return view("admin.projects.edit", compact("project", "type", "technologies"));
     }
 
     /**
@@ -125,8 +138,10 @@ class ProjectController extends Controller
             "description" => "required|string",
             "cover_img" => "file",
             "github_link" => "string",
-            "type_id" => "nullable|exists:types,id"
-            ]);
+            "type_id" => "nullable|exists:types,id",
+            "technology_id" => "array|nullable|exists:technologies,id"
+        ]);
+    
 
 
 
@@ -143,14 +158,19 @@ class ProjectController extends Controller
         }
         $project->update([
             ...$data,
-            "user_id" =>Auth::id(),
-            "cover_img"=>$path ?? $project->cover_img,
+            "user_id" => Auth::id(),
+            "cover_img" => $path ?? $project->cover_img,
 
-        ]);   
-
-
-        return redirect()->route("admin.projects.show",$id);
+        ]);
+        if(isNull($data["technology_id"])){
+            $project->technologies()->detach(); 
+        }else{
+            
+            $project->technologies()->sync($data["technology_id"]);
+        }
         
+
+        return redirect()->route("admin.projects.show", compact("id","project"));
     }
 
     /**
@@ -163,10 +183,10 @@ class ProjectController extends Controller
     {
         $project = Project::findOrFail($id);
 
-      /*   if ($project->cover_img) {
+        if ($project->cover_img) {
             Storage::delete($project->cover_img);
-        } */
-
+        }
+        $project->technologies()->detach();
         $project->delete();
         return redirect()->route("admin.projects.index");
     }
